requirements.txt
Django~=5.0.1
django-modeltranslation==0.18.11